Refactor reclassify_statutory_citations
