+++ tests/test_load_current_murs.py
@@ -25,8 +25,8 @@ def test_parse_statutory_citations_with_reclassifications():
     'url': 'https://api.fdsys.gov/link?collection=uscode&year=mostrecent&link-type=html&title=52&section=30101'}]
 
 def test_parse_statutory_citations_no_reclassifications():
-    assert parse_statutory_citations("30101", 1, 2) == [{'text': '52 U.S.C. 30101',
-    'url': 'https://api.fdsys.gov/link?collection=uscode&year=mostrecent&link-type=html&title=52&section=30101'}]
+    assert parse_statutory_citations("9999", 1, 2) == [{'text': '2 U.S.C. 9999',
+	'url': 'https://api.fdsys.gov/link?collection=uscode&year=mostrecent&link-type=html&title=2&section=9999'}]
 
 def assert_es_index_call(call_args, expected_mur):
     index, doc_type, mur = call_args[0]
